V5 Upload
In V5 the project is running as the teacher wants, receiving and
returning all the correct arguments.
The project is splitter in part1 and 2
Only needs some improvement on the tracking.
In next version, we will be using feature matching between objects, due
to the low frame rate of the cameras.
report and final versions uploaded until 23 of December 2017